task-1.py
This is first task of Python Programming . It is To-Do List . 
A To-Do List application is a useful project that helps users manage and organize their tasks efficiently. This project aims to create a command-line or GUI-based application using Python, allowing users to create, update, and track their to-do list.